Refactor: Move related classes to Reader package; Rename package from home to notelist
