--- app/src/main/java/com/example/user/biblejournal/core/constants/ActivityConstants.kt
@@ -1,7 +1,0 @@
-package com.example.user.biblejournal.core.constants
-
-class ActivityConstants {
-    companion object {
-        val ARG_EDIT_NOTE_ID = "ARG_EDIT_NOTE_ID"
-    }
-}
