--- app/src/main/java/com/example/user/biblejournal/home/viewmodel/HomeViewModel.kt
@@ -1,10 +1,0 @@
-package com.example.user.biblejournal.home.viewmodel
-
-import android.app.Application
-import androidx.lifecycle.AndroidViewModel
-
-class HomeViewModel(app : Application) : AndroidViewModel(app) {
-
-
-
-}
