+++ app/src/main/java/com/example/user/biblejournal/notelist/viewmodel/NoteListViewModel.kt
@@ -1,0 +1,9 @@
+package com.example.user.biblejournal.notelist.viewmodel
+
+import android.app.Application
+import androidx.lifecycle.AndroidViewModel
+
+class NoteListViewModel(app: Application) : AndroidViewModel(app) {
+
+
+}
