+++ app/src/main/java/com/example/user/biblejournal/reader/repository/LocatedVerseAddress.kt
@@ -1,4 +1,4 @@
-package com.example.user.biblejournal.model.data
+package com.example.user.biblejournal.reader.repository
 
 data class LocatedVerseAddress(
         val startIndex: Int,
