+++ app/src/main/java/com/example/user/biblejournal/reader/repository/VerseAddress.kt
@@ -1,4 +1,4 @@
-package com.example.user.biblejournal.model.data
+package com.example.user.biblejournal.reader.repository
 
 data class VerseAddress(
         var book: Int,
