+++ app/src/main/java/com/example/user/biblejournal/writer/view/MyClickableSpan.kt
@@ -3,7 +3,7 @@ package com.example.user.biblejournal.writer.view
 import android.text.TextPaint
 import android.text.style.ClickableSpan
 import android.view.View
-import com.example.user.biblejournal.model.data.VerseAddress
+import com.example.user.biblejournal.reader.repository.VerseAddress
 
 class MyClickableSpan(private val clickListener: ClickableSpanListener, private val verseAddress: VerseAddress) : ClickableSpan() {
