+++ app/src/main/res/layout/fragment_note_list.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#272822"
-    tools:context=".home.view.NoteListFragment">
+    tools:context=".notelist.view.NoteListFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rvNoteList"
